Add most types, ensure window resizes correctly
All but one declaration (which appeared to be unsatisfiable in
Minigrace) has been fully typed. The check for window resizing was not
perfect, and so has been adjusted to just keep resizing (a no-op once it
has been resized correctly) for half a second, which is a reasonable
time to wait.