Tie the auto-resizing of a window to the window
The interval and timeout involved in auto-resizing a window was spawned
by the original window, not the window they were resizing. The outcome
was that when the child window was closed, the interval was still trying
to resize the window, which resulted in errors in some browsers. Now the
interval and timeout are spawned by the child, and stop when the child
is closed.