Ensure select boxes implement the Choice type
The selectBox class was missing the required methods from the Labeled
type, which meant its instances did not satisfy the Choice type. Now
select boxes can be labelled, which produces an initial option with an
empty value. Currently setting the label on an initially unlabelled
select box does not add the label to the box.