Use page* instead of client* for event points
The calculation of event points used the clientX and clientY properties
of the event object, which is relative to the top-left corner of the
client, which means it did not take into account scrolling when
comparing it against the element's offset values. Now the calculation
uses pageX and pageY, will will provide the correct values if the page
is scrolled.