Reverts some of Kim's edits to Objectdraw.
Kim is using an old version of the compiler.
He is working around compiler and dom module bugs that have been
fixed, and in the process creating a version of objectdraw that
won't compile with the current compiler.  This is preventing
us from pushing the new compiler to the web editor.